fixed overflow

+++ mempool/src/topologies/impls.rs
@@ -178,7 +178,8 @@ impl Topology for BinomialTreeTopology {
                 base >>= 1;
             }
         } else {
-            let mut tmp = (self.my_index - index).rem_euclid(self.peers.len());
+            let mut tmp =
+                (self.my_index as i32 - index as i32).rem_euclid(self.peers.len() as i32) as usize;
             let fixed_index = tmp;
             while tmp != 0 && tmp % 2 == 0 {
                 tmp >>= 1;
